Refactor: Move tree parameter tests from RendererTest.cs their own file
